Additional methods to read an XML version of the PSMV

Additional methods to read an XML version of the PSMV

Usage

# S3 method for numeric
psmv_xml_get(from, ...)

# S3 method for Date
psmv_xml_get(from, ...)

Arguments

from

A number giving a year starting from 2011 up to the current year, or one of the dates in psmv_xml_dates.

...

Currently not used

Value

An object inheriting from 'psmv_xml', 'xml_document', 'xml_node'

Examples

# \dontrun{
# The following only works if you have a collection of zipped XML dumps in a directory
# specified in the environment variable R_psmvhist_idir
psmv_2015 <- psmv_xml_get(2015)
#> ℹ Reading XML for 2015-01-06
print(psmv_2015)
#> {xml_document}
#> <PublicationData>
#>  [1] <Products numberOfProducts="2096">\n  <Product terminationReason="" isSa ...
#>  [2] <Parallelimports numberOfParallelimports="1255">\n  <Parallelimport pack ...
#>  [3] <MetaData numberOfRows="11" name="ApplicationArea">\n  <Detail primaryKe ...
#>  [4] <MetaData numberOfRows="760" name="ApplicationComment">\n  <Detail prima ...
#>  [5] <MetaData numberOfRows="6" name="CultureForm">\n  <Detail primaryKey="64 ...
#>  [6] <MetaData numberOfRows="328" name="Culture">\n  <Detail primaryKey="1139 ...
#>  [7] <MetaData numberOfRows="56" name="FormulationCode">\n  <Detail primaryKe ...
#>  [8] <MetaData numberOfRows="105" name="Measure">\n  <Detail primaryKey="6568 ...
#>  [9] <MetaData numberOfRows="1316" name="Obligation">\n  <Detail primaryKey=" ...
#> [10] <MetaData numberOfRows="580" name="Pest">\n  <Detail primaryKey="11001"> ...
#> [11] <MetaData numberOfRows="31" name="ProductCategory">\n  <Detail primaryKe ...
#> [12] <MetaData numberOfRows="3" name="TimeMeasure">\n  <Detail primaryKey="73 ...
#> [13] <MetaData numberOfRows="495" name="Substance">\n  <Detail iupacName="1-n ...
#> [14] <MetaData numberOfRows="269" name="PermissionHolder">\n  <PermissionHold ...
#> [15] <MetaData numberOfRows="225" name="City">\n  <Detail primaryKey="3285">\ ...
#> [16] <MetaData numberOfRows="15" name="Country">\n  <Detail primaryKey="6161" ...
#> [17] <MetaData numberOfRows="80" name="CodeR">\n  <Detail primaryKey="5978">\ ...
#> [18] <MetaData numberOfRows="91" name="CodeS">\n  <Detail primaryKey="6099">\ ...
#> [19] <MetaData numberOfRows="15" name="DangerSymbol">\n  <Detail primaryKey=" ...
#> [20] <MetaData numberOfRows="2" name="SignalWords">\n  <Detail primaryKey="79 ...
#> ...
class(psmv_2015)
#> [1] "psmv_xml"     "xml_document" "xml_node"    

# This is the method for dates used behind the scenes
psmv_2017 <- psmv_xml_get(as.Date("2017-01-13"))
#> ℹ Reading XML for 2017-01-13
print(psmv_2017)
#> {xml_document}
#> <PublicationData>
#>  [1] <Products numberOfProducts="2269">\n  <Product terminationReason="TERMIN ...
#>  [2] <Parallelimports numberOfParallelimports="1540">\n  <Parallelimport pack ...
#>  [3] <MetaData numberOfRows="10" name="ApplicationArea">\n  <Detail primaryKe ...
#>  [4] <MetaData numberOfRows="755" name="ApplicationComment">\n  <Detail prima ...
#>  [5] <MetaData numberOfRows="5" name="CultureForm">\n  <Detail primaryKey="64 ...
#>  [6] <MetaData numberOfRows="337" name="Culture">\n  <Detail primaryKey="1139 ...
#>  [7] <MetaData numberOfRows="55" name="FormulationCode">\n  <Detail primaryKe ...
#>  [8] <MetaData numberOfRows="117" name="Measure">\n  <Detail primaryKey="6568 ...
#>  [9] <MetaData numberOfRows="1401" name="Obligation">\n  <Detail primaryKey=" ...
#> [10] <MetaData numberOfRows="574" name="Pest">\n  <Detail primaryKey="11956"> ...
#> [11] <MetaData numberOfRows="27" name="ProductCategory">\n  <Detail primaryKe ...
#> [12] <MetaData numberOfRows="3" name="TimeMeasure">\n  <Detail primaryKey="73 ...
#> [13] <MetaData numberOfRows="534" name="Substance">\n  <Detail iupacName="1-n ...
#> [14] <MetaData numberOfRows="269" name="PermissionHolder">\n  <PermissionHold ...
#> [15] <MetaData numberOfRows="225" name="City">\n  <Detail primaryKey="3285">\ ...
#> [16] <MetaData numberOfRows="17" name="Country">\n  <Detail primaryKey="6161" ...
#> [17] <MetaData numberOfRows="106" name="CodeR">\n  <Detail primaryKey="5978"> ...
#> [18] <MetaData numberOfRows="97" name="CodeS">\n  <Detail primaryKey="13344"> ...
#> [19] <MetaData numberOfRows="16" name="DangerSymbol">\n  <Detail primaryKey=" ...
#> [20] <MetaData numberOfRows="2" name="SignalWords">\n  <Detail primaryKey="79 ...
#> ...
# }